% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{addIgG}
\alias{addIgG}
\title{Add IgG distribution}
\usage{
addIgG(data, IgG, name)
}
\arguments{
\item{data}{An object of class \code{CygnusObject}}

\item{IgG}{A vector for IgG distribution}

\item{name}{A string specifying the name of IgG distribution to be saved in \code{CygnusObject}}
}
\value{
The updated CygnusObject with IgG distribution added
}
\description{
Add IgG distribution
}
