% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateCygnus}
\alias{CreateCygnus}
\title{Create a Cygnus Object}
\usage{
CreateCygnus(data.path, markers_col = NULL, meta_col = NULL)
}
\arguments{
\item{data.path}{Character string specifying the path to the data file.}

\item{markers_col}{A vector of column names or indices representing markers.}

\item{meta_col}{A vector of column names or indices representing metadata.}
}
\value{
An object of class \code{CygnusObject}.
}
\description{
This function creates a Cygnus Object from the AIVIA output file format.
If no columns are provided, a Shiny app will be launched to allow
interactive selection of columns.
}
