% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{CygnusObject-class}
\alias{CygnusObject-class}
\alias{CygnusObject}
\title{The Cygnus Object Class}
\description{
The CygnusObject class is a data storage class that stores expression data
and other related information needed for the standard Cygnus workflow.
}
\section{Slots}{

\describe{
\item{\code{matrices}}{List of matrices}

\item{\code{ev_meta}}{Metadata for each sample}

\item{\code{markers_meta}}{Metadata for each marker}

\item{\code{marker_analysis}}{Results from marker characterization analyses}

\item{\code{dim_red}}{Coordinates from dimensionality reduction analysis}
}}

