% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{getColocalizedMarkers}
\alias{getColocalizedMarkers}
\title{Generate Marker Colocalization Deviation and Statistics}
\usage{
getColocalizedMarkers(
  data,
  matrix_name = "binary_exp_matrix",
  iteration_no = 100
)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing the binary expression matrix.}

\item{matrix_name}{The matrix within the data object that contains the binary expression matrix. The default is set to binary_exp_mat}

\item{iteration_no}{The number of iterations for simulation}
}
\value{
A dataframe of deviation and statistics of each intersection
}
\description{
Generate Marker Colocalization Deviation and Statistics
}
