% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{markRelevantMarkersUI}
\alias{markRelevantMarkersUI}
\title{Mark Relevant Markers Using User Interface}
\usage{
markRelevantMarkersUI(data)
}
\arguments{
\item{data}{An object of class \code{CygnusObject}.}

\item{relevant_markers}{A character vector of marker names to be marked as relevant. If NULL, no markers are marked as relevant.}
}
\value{
The updated \code{CygnusObject} with relevant markers marked.
}
\description{
This function updates the markers_meta slot in a CygnusObject to mark specified markers as relevant.
The markers_meta must be a data frame with a 'marker' column. The function adds a 'relevant' list.
}
