% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{nnormalizeByPanEV}
\alias{nnormalizeByPanEV}
\title{Normalize Expression Data by Pan-EV Marker}
\usage{
nnormalizeByPanEV(data, pan_ev_marker)
}
\arguments{
\item{data}{An object of class \code{CygnusObject}.}

\item{pan_ev_marker}{Character string specifying the name of the pan-EV marker.}
}
\value{
The updated CygnusObject with normalized expression matrix.
}
\description{
This method normalizes the expression matrix in the CygnusObject by dividing
the expression values of all markers by the expression values of a specified
pan-EV marker.
}
