% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{plotAvgHeatmap}
\alias{plotAvgHeatmap}
\title{Plot Average Expression Heatmap by Group}
\usage{
plotAvgHeatmap(
  data,
  group_column,
  clustering_distance = "euclidean",
  colors = rev((grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 11,
    "RdYlBu")))(100)),
  fontsize = 8,
  scale = "column",
  cluster_rows = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing an expression matrix and metadata.}

\item{group_column}{Character string specifying the metadata column to group by.}

\item{clustering_distance}{Character string specifying the distance metric for clustering columns. Default is "euclidean".}

\item{colors}{A color palette to use for the heatmap. Default is a reversed "RdYlBu" palette from RColorBrewer.}

\item{fontsize}{Numeric value specifying the font size for heatmap text. Default is 8.}

\item{scale}{Character string specifying whether the data should be scaled by 'row', 'column', or 'none'. Default is 'column'.}

\item{cluster_rows}{Logical value indicating whether to cluster rows in the heatmap. Default is FALSE.}

\item{na.rm}{Logical value indicating whether to remove NA values before calculating averages. Default is TRUE.}
}
\value{
A heatmap plot showing average marker expressions grouped by the specified metadata column.
}
\description{
This function generates a heatmap of average marker expressions, grouped by a specified metadata column.
The heatmap allows customization of clustering distance, color palette, font size, and scaling method.
}
