% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotElbowPlot}
\alias{plotElbowPlot}
\title{Plot Elbow Plot for PCA}
\usage{
plotElbowPlot(data, ...)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing PCA results.}

\item{...}{Additional arguments to pass to the plot function.}
}
\value{
An elbow plot showing the variance explained by each principal component.
}
\description{
Plot Elbow Plot for PCA
}
