% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{plotLine}
\alias{plotLine}
\title{Visualize Linegraph}
\usage{
plotLine(
  data,
  x_axis,
  plot_markers = "ALL",
  fold_change = T,
  matrix_name = "Raw_Score",
  order_x_axis = NULL
)
}
\description{
Makes sense for temporal data!
}
