% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotPCA}
\alias{plotPCA}
\title{Visualize PCA Results}
\usage{
plotPCA(data, plot_3d = FALSE, color_by = NULL, marker_size = 5, ...)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing PCA results.}

\item{plot_3d}{Logical value indicating whether to plot in 3D. Default is FALSE.}

\item{color_by}{Character string specifying metadata column to color by. Default is NULL (no coloring).}

\item{marker_size}{Size of the points}

\item{...}{Additional arguments for the plot function}
}
\value{
A plot of the PCA results.
}
\description{
Visualize PCA Results
}
