% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotTSNE}
\alias{plotTSNE}
\title{Visualize t-SNE Results}
\usage{
plotTSNE(data, plot_3d = FALSE, color_by = NULL, marker_size = 1, ...)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing t-SNE results.}

\item{plot_3d}{Logical value indicating whether to plot in 3D. Default is FALSE.}

\item{color_by}{Character string specifying metadata column to color by. Default is NULL (no coloring).}

\item{marker_size}{Numeric value specifying the size of the markers in the plot. Default is 1.}
}
\value{
A plot of the t-SNE results.
}
\description{
Visualize t-SNE Results
}
