% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Visualize UMAP Results}
\usage{
plotUMAP(data, plot_3d = FALSE, color_by = NULL)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing UMAP results.}

\item{plot_3d}{Logical value indicating whether to plot in 3D. Default is FALSE.}

\item{color_by}{Character string specifying metadata column to color by. Default is NULL (no coloring).}
}
\value{
A plot of the UMAP results.
}
\description{
Visualize UMAP Results
}
