% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{plotUpset}
\alias{plotUpset}
\title{Generate an UpSet Plot from Colocalization Statistics}
\usage{
plotUpset(
  data,
  colocal_data,
  matrix_name = "binary_exp_mat",
  threshold_count = 1,
  min_degree_setting = 1
)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing the binary expression matrix.}

\item{colocal_data}{A data frame of deviations and statistics from running getColocalizedMarkers}

\item{matrix_name}{The matrix within the data object that contains the binary expression matrix. The default is set to binary_exp_mat}

\item{threshold_count}{An integer indicating the minimum number of elements in an intersection}

\item{min_degree_setting}{An integer indicating the minimum degree of intersection for plotting}
}
\description{
Generate an UpSet Plot from Colocalization Statistics
}
