% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Generate an Volcano Plot from Colocalization Statistics}
\usage{
plotVolcano(colocal_data, sig_threshold = 0.05)
}
\arguments{
\item{colocal_data}{A data frame of deviations and statistics from running getColocalizedMarkers}

\item{sig_threshold}{The threshold for significance}
}
\description{
Generate an Volcano Plot from Colocalization Statistics
}
