% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{runPCA}
\alias{runPCA}
\title{Perform PCA on a Specified Matrix and Store in Dimensionality Reduction Slot}
\usage{
runPCA(
  data,
  matrix_name = "Raw_Score",
  num_components = 50,
  use_relevant = TRUE
)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing an expression matrix.}

\item{matrix_name}{Character string specifying the name of the matrix to use. Default is "Raw_Score".}

\item{num_components}{Numeric value specifying the number of principal components to compute. Default is 50.}

\item{use_relevant}{Logical indicating whether to use only relevant markers. Default is TRUE.}
}
\value{
The updated CygnusObject with PCA coordinates stored in the 'dim_red' slot.
}
\description{
Perform PCA on a Specified Matrix and Store in Dimensionality Reduction Slot
}
