% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{scaleExpressionMatrix}
\alias{scaleExpressionMatrix}
\title{Scale Expression Matrix by Maximum Value}
\usage{
scaleExpressionMatrix(data, matrix_name = "scaled_exp_matrix")
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing an expression matrix.}

\item{matrix_name}{Character string specifying the name for the new scaled matrix layer. Default is "scaled_exp_matrix".}
}
\value{
The data object with an additional layer containing the scaled expression matrix.
}
\description{
This function scales the expression matrix by dividing each marker's values by the maximum value in that marker.
The scaled matrix is added as an additional layer in the data object.
}
