% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{specifyColumns}
\alias{specifyColumns}
\title{Create Cygnus by Launching Shiny UI for Column Selection}
\usage{
specifyColumns(data.path)
}
\arguments{
\item{data.path}{Character string specifying the path to the data file.}
}
\value{
A list containing selected columns for markers and metadata.
}
\description{
This function launches a Shiny app that allows users to select columns for markers
and metadata interactively. It then returns the selected columns.
}
