% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExploration.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plot Distributions of Selected Markers}
\usage{
plotDistribution(obj, plot_markers = "ALL", matrix = "Raw_Score")
}
\arguments{
\item{obj}{An object of class \code{CygnusObject} containing the expression matrix.}

\item{plot_markers}{A vector of marker names to plot. If set to "ALL", distributions for all markers will be plotted. Default is "ALL".}

\item{matrix}{Character string specifying the name of the matrix to use. Default is "Raw_Score".}
}
\value{
A series of histogram plots, each representing the distribution of expression levels for the specified markers.
}
\description{
This function plots the distribution of expression levels for selected markers in an expression matrix.
The user can specify a subset of markers to plot, or plot all markers by default.
}
