% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-SNE on a Specified Matrix and Store in Dimensionality Reduction Slot}
\usage{
runTSNE(
  data,
  matrix_name = "Raw_Score",
  use_relevant = TRUE,
  dims = 2,
  use_pcs = FALSE,
  ...
)
}
\arguments{
\item{data}{An object of class \code{CygnusObject} containing an expression matrix.}

\item{matrix_name}{Character string specifying the name of the matrix to use. Default is "Raw_Score".}

\item{use_relevant}{Logical indicating whether to use only relevant markers. Default is TRUE.}

\item{dims}{Numeric value specifying the number of dimensions for t-SNE. Default is 2.}

\item{use_pcs}{Logical indicating whether to use PCA results as input. Default is FALSE.}

\item{...}{Additional arguments to pass to the Rtsne function.}
}
\value{
The updated CygnusObject with t-SNE coordinates stored in the 'dim_red' slot.
}
\description{
Perform t-SNE on a Specified Matrix and Store in Dimensionality Reduction Slot
}
